# PctBackend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion Heures UVCI API",
        "description": "Documentation de l'API de gestion des heures",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Connexion utilisateur",
                "operationId": "bae68d1c55329ecb6e41f48285840c8f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password"
                                ],
                                "properties": {
                                    "login": {
                                        "type": "string",
                                        "example": "admin@uvci.edu.ci"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Identifiants invalides"
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Déconnexion",
                "operationId": "c5cdd87fdbf9d0f95dede875bd023dec",
                "responses": {
                    "200": {
                        "description": "Déconnecté avec succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Utilisateur connecté",
                "operationId": "13b43df400e475a157538f73617b8c3d",
                "responses": {
                    "200": {
                        "description": "Données de l'utilisateur connecté"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/cours": {
            "get": {
                "tags": [
                    "Cours"
                ],
                "summary": "Liste des cours",
                "operationId": "d6c49484ccc20ce1867c367f94dd7c3a",
                "parameters": [
                    {
                        "name": "niveau",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des cours"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cours"
                ],
                "summary": "Créer un cours",
                "operationId": "cd05524e774d700c4d4e40115be7170b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "intitule_ecue",
                                    "niveau",
                                    "semestre",
                                    "credit_ecue",
                                    "charge_horaire_annuel"
                                ],
                                "properties": {
                                    "intitule_ecue": {
                                        "type": "string",
                                        "example": "Algorithmes avancés"
                                    },
                                    "niveau": {
                                        "type": "string",
                                        "enum": [
                                            "L1",
                                            "L2",
                                            "L3",
                                            "M1",
                                            "M2"
                                        ]
                                    },
                                    "semestre": {
                                        "type": "string",
                                        "example": "S3"
                                    },
                                    "credit_ecue": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "code_specialite": {
                                        "type": "string",
                                        "example": "INFO"
                                    },
                                    "charge_horaire_annuel": {
                                        "type": "number",
                                        "example": 45
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cours créé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/departements": {
            "get": {
                "tags": [
                    "Départements"
                ],
                "summary": "Liste de tous les départements",
                "operationId": "fb736a0bd46ca03123e83c6d7d0342b9",
                "responses": {
                    "200": {
                        "description": "Liste des départements"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Départements"
                ],
                "summary": "Créer un département",
                "operationId": "822e20de313a1b7d94c5c0d0ac904cfb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom_departement"
                                ],
                                "properties": {
                                    "nom_departement": {
                                        "type": "string",
                                        "example": "Informatique"
                                    },
                                    "responsable": {
                                        "type": "string",
                                        "example": "Dr. Koffi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Département créé"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/departements/{id}": {
            "get": {
                "tags": [
                    "Départements"
                ],
                "summary": "Détail d'un département avec ses enseignants",
                "operationId": "f3c2396b4460f01dca7dafce020494ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail département"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Départements"
                ],
                "summary": "Modifier un département",
                "operationId": "e32afea528656ac0c7b0c9f327d75443",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom_departement": {
                                        "type": "string"
                                    },
                                    "responsable": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Département modifié"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Départements"
                ],
                "summary": "Supprimer un département",
                "operationId": "5ad105f5d17f17eb76b4bad5d8d21aa4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Département supprimé"
                    },
                    "422": {
                        "description": "Impossible : des enseignants sont rattachés à ce département"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}